better support for blank field names when comparing them: SERVER-2651
When passing a blank field name as the second parameter to compareDottedFieldNames, the function
was returning LEFT_SUBFIELD.  By reordering the logic so that we only check for subfields after
the first lexicographic compare, we now return the correct answer.
The code change will not affect other cases because the subfield tests test to see if we've reached
the end of the strings with our parsing.  Before the first iteration of the loop, this situation will
never be true unless one of the field names is blank.