SERVER-4312 implement kill/yank in mongo shell
A single kill buffer is now working in the mongo shell.  All of
the commands that are supposed to add text to the kill buffer now
do so, and accumulate text properly with repeated kills in either
direction (to the left or to the right) or both.  No Meta-Y (rotate
kill buffer, "yank-pop") feature in this checkin, so it's not a
kill ring yet.