fix socket timeouts on Windows
It turns out that SO_RCVTIMEO and SO_SNDTIMEO socket options on Windows
do not take a struct timeval like they do on most Unixes.  Instead, they
take a DWORD which is the number of milliseconds.
Most sockets we create do not have a timeout, but for the ones that do, we were
setting the timeout to 30 ms instead of 30 seconds.  This is why dumprestore9.js
unit test was intermittantly failing.