SERVER-4312 rewrite tab completion to work like readline
Replace the old logic for rotating through completion possibilities
with logic that mimics readline's behavior.  We now only complete
the portion that is unambiguous (the same in all possibilities) and
show a list of matches if tab is hit again after a failure to
complete a match.  If the number of matches is over 100, we ask
before displaying the list.  Still to come -- paging code to pause
after displaying a screen's worth of matches.