SERVER-5073 fix failing test in 32-bit Windows
Use the "force" option in the Windows version's kill_wrapper()
routine (called by the StopMongoProgram() and related shell
commands) so we can shut down a replica set primary after we have
shut down the secondaries.  Increase the timeout before using the
Windows TerminateProcess() API from 30 seconds to 60 seconds to
reduce the window for a race.