Promote C++ batch query execution functions to DBClientBase.
There's no good reason I can think of why the batch-processing forms of query()
shouldn't be available to all implementations of DBClientBase, and simply
optimized for DBClientConnection when exhaust-mode is available on the
connection.

This change facilitates another optimization, coming soon, which is that
DBDirectClient's batch query form can scan all rows rather than issuing a find
query when the query parameters are appropriate.  This allows for a clever
resolution of SERVER-4726, which is really about an unwanted log message.