Roll back most of the following two patches:
c12bfcd76a1c101d16416f5d26dbf6382d2289d0:  qlock can ask if in greedy mode or not
e97c465df6f46fdb77e04aac5c5e400d020ea093: if in ThreadSpanningOp don't profile

The problem is that isGreedSuspended cannot be called with the qlock's internal
mutex already held.  However, a caller not holding that lock already can't
know that by the time the function returns, its return value is still valid.
Another thread may change it in parallel.  The result is that the attempt
to use it to stop profiling when thread-spanning ops are enabled only appears
to provide that protection.