SERVER-5548 Relax timing in remove1.js
It seems to take almost exactly 30 seconds to cycle a primary
up or down on the 32-bit Windows buildbot and the test times
out after 30 seconds so this is too flaky.  Change the timeout
from 30 seconds to 60 seconds to buy some breathing room.  Also,
the comments and messages displayed by this test weren't aligned
with changes that had been made to it, making the log hard to
follow: fix up the comments and log messages to match what the
test is actually doing.