Limit exitCleanly() and ntservice module to mongos and mongod.
The notions of exitCleanly() versus dbexit() vs _exit() in the scope of mongod
shutdown need to be reconsidered.  This patch just attempts to limit the scope
of the issue to mongos and mongod.  It would be better to limit it to just
mongod, but the implementation of the ntservice behavior doesn't currently
support that.

SERVER-5520, SERVER-5702