SERVER-5867 Use CRITICAL_SECTION for non-SRW Windows locks
Also related to SERVER-2942.  The non-2008plus Windows builds was using
the Microsoft Concurrency Runtime for the SimpleMutex class, but this
interacts badly with the Windows memory hooking code added for SERVER-2942.
A call to VirtualAlloc during thread creation would take out the SimpleMutex
for RemapLock, which (through the Concurrency Runtime) make another call
into VirtualAlloc, deadlocking the thread.  I changed SimpleMutex to use
the Win32 CRITICAL_SECTION APIs, which do not allocate memory when taking
the critical section.