SERVER-6718 new parameter to control replication index prefetching behavior
New parameter replIndexPrefetch.
Settable via the command line or use setParameter() command on a running server.
Only valid on a server started with --replSet parameter.
This allows a user to disable the prefetching of all index pages prior to the application
of a replicated op on a secondary.
For most use cases, prefetching all indexes for a given op will provide good performance.
For use cases that do many in-place updates to documents in a collection with many indexes
that are unaffected by such updates, using the '_id_only' setting may provide better performance
than the 'all' setting.  The '_id_only' setting only prefetches index pages for the _id index,
which will always be used to find the document needed to be updated.
The 'none' setting will be used for debugging only; there should be no production use cases for
this setting.
getParameter() supports getting the current value of this parameter.