SERVER-7617 Enter and Exit our scope's Isolate in V8
Added Scope::enter() and exit() member functions, only doing something
in V8 builds, to enter and exit the Isolate of the Scope we are using.
Call these functions when fetching a Scope from the pool/ScopeCache and
when returning one to the pool or deleting one.  The Isolate is the actual
"state" container for V8, so many things don't work properly if this isn't
set up.  There may be other situations where we need this, but this fixes
(at least) this particular case.