SERVER-12616 Presume that the localhost exception does not apply in mongos when configs unreachable.
If the config servers become unreachable by a mongos server in systems with access control enabled,
that mongos must assume that there are user documents in the config server and so the localhost
exception does not apply. Otherwise, one needs only block access to the config servers to bypass
access controls.  While the mongos won't be able to do metadata operations, it will be able to
read and write arbitrary data, when directed to do so by a client on the localhost interface.