SERVER-14538 _appendUserInfo: don't save addr pointing to temporary
Before 1fb80ae7, the return value of CurOp::getNS() had a lifetime of
the CurOp object.   As of 1fb80ae7, CurOp::getNS() returns a
temporary.

_appendUserInfo() initialized a StringData from the result of
CurOp::getNS(), which was poor style but technically valid before
1fb80ae7.  As of 1fb80ae7, the buffer being pointed to by the
StringData is destroyed along with the temporary, creating a
use-after-free at introspect.cpp:70 when the StringData is read.