SERVER-15532 Fix assumptions made by geo_s2cursorlimitskip.js test.
For storage engines that are transactional, the documents that are
inserted after the creation of a cursor need not be visible to the
cursor. This means that the set of points inserted after the query
is issued may not be found, depending on the storage engine.

Note that this test can no longer be run in parallel because we
assert a specific number of query and getmore operations have
occurred.

Signed-off-by: Dan Pasette <dan@mongodb.com>