SERVER-16376 Make nodes in jstests/gle/get_last_error.js choose their sync source deterministically.
The test expects w:majority writes to happen faster than a heartbeat interval when a single
non-primary node fails.  If, however, that failed node is the intermediate in a replication
chain, there can be a multi-second lag until the other nodes recognize the failure.  Change
the test to disallow chaining, because it is not the subject of the test.