SERVER-16651 Avoid global lock acquisition in tryToGoLiveAsASecondary() if a primary or secondary.
Only nodes in state RECOVERING will attempt a state transition in
tryToGoLiveAsASecondary(), and we have a lock-free way to check if a node is in
one of those two states.  The danger of asking the question, "is this node in
either primary or secondary state?" without holding the global lock is that the
answer might change at any time.  However, only tryToGoLiveAsASecondary() causes
it to transition from "no" to "yes", and if it transitions from "yes" to "no",
we know that tryToGoLiveAsASecondary() will be called again promptly.