SERVER-18756 Write all executions to the report.json file.
Makes it so that a link to the logs for every test that ran is present
in the sidebar of the Evergreen UI. Multiple entries will appear for
the same test file when --repeat is used.

This also fixes an issue where the number of tests skipped would be
incorrect if the same test file was included multiple times in the
"roots" key.