SERVER-19000 Fix thread_pool_test.
While joining a ThreadPool, the value of the "numThreads" stats value is locked
at the number of threads that existed when the join operation began, even if one
or more of those threads have since exited. This is because the numThreads value
is computed as the number of threads in the _threads container, which are the
threads that will be joined individually during a pool join operation.

As such, it is not correct to make assertions about the value of numThreads
while ThreadPool::join is running, as it is in DestructionDuringDoubleJoins.