SERVER-21193 mongos claims that config db is empty on CSRS cluster
dbStats don't support readConcern since database metadata in the server are not synchronized with the snapshots. This makes the listDatabase command in mongos to send listDatabase instead with primaryPreferred and without the readConcern option.

This also changed the listDatabase behavior to not include the config or admin in the result if they don't actually exist.